Update (src/run.py & common/utils.py): moved method log_hyperparameters from utils.py to run.py, and remove import pytorch_lightning Update (Dockerfile): add environment variables for wandb
