+++ Dockerfile
@@ -1,11 +1,16 @@
-FROM amazon/aws-lambda-python
+FROM public.ecr.aws/lambda/python:3.8
 
 ARG AWS_ACCESS_KEY_ID
 ARG AWS_SECRET_ACCESS_KEY
 ARG MODEL_DIR=./models
 ARG WANDB_DIR=./wandb_dir
+ARG WANDB_CONFIG_DIR=./wandb_config_dir
+ARG WANDB_CACHE_DIR=./wandb_cache_dir
 RUN mkdir $MODEL_DIR
 RUN mkdir $WANDB_DIR
+RUN mkdir $WANDB_CONFIG_DIR
+RUN mkdir $WANDB_CACHE_DIR
+
 
 ENV TRANSFORMERS_CACHE=$MODEL_DIR \
     TRANSFORMERS_VERBOSITY=error
@@ -23,6 +28,8 @@ RUN pip install -r requirements.txt --no-cache-dir
 COPY ./ ./
 ENV PROJECT_ROOT=./
 ENV WANDB_DIR=$WANDB_DIR
+ENV WANDB_CONFIG_DIR=$WANDB_CONFIG_DIR
+ENV WANDB_CACHE_DIR=$WANDB_CACHE_DIR
 
 ENV PYTHONPATH "${PYTHONPATH}:./"
 ENV LC_ALL=C.UTF-8
@@ -31,8 +38,12 @@ ENV LANG=C.UTF-8
 # pulling the trained model
 RUN dvc pull models/model.onnx.dvc
 
-RUN ls
 RUN chmod -R 0755 $WANDB_DIR
+RUN chmod -R 0755 $WANDB_CONFIG_DIR
+RUN chmod -R 0755 $WANDB_CACHE_DIR
+
+RUN chmod -R 0755 ./
+
 RUN python -m src.lambda_handler
 RUN chmod -R 0755 $MODEL_DIR
 CMD ["src.lambda_handler.lambda_handler"]
